+++ src/Ext/Rules/Body.h
@@ -282,6 +282,8 @@ public:
 		int TintColorForceShield;
 		int TintColorBerserk;
 
+		std::vector<std::string> EVAIndexList;
+
 		Valueable<bool> InfantryAutoDeploy;
 
 		Nullable<bool> BattlePoints;
@@ -507,6 +509,8 @@ public:
 			, AttackMove_IgnoreWeaponCheck { false }
 			, AttackMove_StopWhenTargetAcquired { }
 
+			, EVAIndexList {}
+
 			, Parasite_GrappleAnim {}
 			, InfantryAutoDeploy { false }
 
@@ -530,6 +534,7 @@ public:
 		virtual void SaveToStream(PhobosStreamWriter& Stm) override;
 
 		void ReplaceVoxelLightSources();
+		void LoadEvaVoices();
 
 	private:
 		template <typename T>