+++ docs/Whats-New.md
@@ -466,6 +466,7 @@ New:
 - Health bar permanently displayed (by FlyStar)
 - [`IsSimpleDeployer` facing customization & directional deploy animations](Fixed-or-Improved-Logics.md#issimpleDeployer-facing-and-animation-customization) (by Starkku)
 - Battle Points economy for super weapons (by FS-21)
+- New EVA voice after deploying a building (by FS-21)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)